# api/core-api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenTreasury Core API
  version: 0.1.0
  description: >-
    Core treasury transaction API for OpenTreasury.


    Conventions: every error response uses the ErrorResponse envelope. Internal
    failures return a sanitized `internal server error` message (details are
    logged server-side with the request ID, never returned). All responses
    carry an `X-Request-ID` header — clients may supply their own. `limit`
    parameters are validated strictly: values outside 1..100 are rejected with
    400 rather than silently clamped. Request bodies above 1 MiB are rejected
    with 413. Cross-origin requests are allowed only for origins configured via
    `OPENTREASURY_ALLOWED_ORIGINS` (methods GET, POST, OPTIONS; header
    Content-Type).
servers:
  - url: http://localhost:8080
paths:
...
  /v1/accounts:
    get:
      operationId: listAccounts
      summary: List accounts of the active chart of accounts
      parameters:
        - name: accountType
          in: query
          required: false
          schema:
            type: string
            enum:
              - ASSET
              - LIABILITY
              - NET_WORTH
              - REVENUE
              - EXPENSE
          description: Filter accounts by classification type.
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Accounts of the active chart version in code order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAccountsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          description: Account persistence is not available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/journal-entries:
    get:
      operationId: listJournalEntries
      summary: List posted journal entries
      parameters:
        - name: institutionId
          in: query
          required: false
          schema: { type: string, minLength: 1 }
        - name: fiscalYear
          in: query
          required: false
          schema: { type: integer, minimum: 1 }
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [POSTED, REVERSED] }
        - $ref: "#/components/parameters/dateFrom"
        - $ref: "#/components/parameters/dateTo"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Matching journal entries with their lines.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListJournalEntriesResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: Journal persistence is not available.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      operationId: postJournalEntry
      summary: Post a balanced double-entry journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JournalEntryInput" }
      responses:
        "201":
          description: Entry posted; balances updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateEntryResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: An entry with the same id already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422":
          description: The referenced institution does not exist.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: Journal persistence is not available.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/balances:
    get:
      operationId: listBalances
      summary: List materialized account balances
      parameters:
        - name: institutionId
          in: query
          required: false
          schema: { type: string, minLength: 1 }
        - name: accountCode
          in: query
          required: false
          schema: { type: string, minLength: 1 }
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Balances per institution, account, and currency.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ListBalancesResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: Balance persistence is not available.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /v1/audit-events:
    get:
      operationId: listAuditEvents
      summary: List treasury audit events
      parameters:
        - name: institutionId
          in: query
          required: false
          schema:
            type: string
            minLength: 1
          description: Filter audit events by institution identifier.
        - $ref: "#/components/parameters/dateFrom"
        - $ref: "#/components/parameters/dateTo"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Matching audit events ordered by most recent occurrence.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAuditEventsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          description: Audit event persistence is not available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/institutions:
    get:
      operationId: listInstitutions
      summary: List treasury institutions
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
          description: Filter institutions by status.
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Treasury institutions ordered by name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListInstitutionsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          description: Institution persistence is not available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/transactions/validate:
    post:
      operationId: validateTransaction
      summary: Validate a treasury transaction without persisting it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreasuryTransaction"
      responses:
        "200":
          description: Transaction is valid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
...
  /v1/reconciliation:
    get:
      operationId: listReconciliation
      summary: Reconcile staged source records against posted journal entries
      description: >-
        Per source system and calendar month: how many records arrived in
        staging, how many posted or were quarantined, and whether the staged
        debit total matches what was actually posted to the journal.
      parameters:
        - name: sourceSystem
          in: query
          required: false
          schema:
            type: string
            minLength: 1
          description: Filter by originating source system.
        - name: fiscalYear
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Filter by fiscal year.
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Reconciliation rows for the selected filters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListReconciliationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "503":
          description: Reconciliation persistence is not available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/entries/{id}/proof:
    get:
      operationId: getEntryProof
      summary: Get the tamper-evidence proof for a journal entry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Anchor receipt + Merkle inclusion proof.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EntryProof" }
        "404":
          description: The entry has not been anchored yet.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  parameters:
    page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-based page number.
    pageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 15
      description: Rows per page.
    limit:
      name: limit
      in: query
      required: false
      deprecated: true
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Deprecated alias for pageSize.
    dateFrom:
      name: dateFrom
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Inclusive lower date bound.
    dateTo:
      name: dateTo
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Inclusive upper date bound. Must not precede dateFrom.
  responses:
...
    BadRequest:
      description: Request payload is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: >-
        Unexpected server failure. The message is always the sanitized string
        `internal server error`; diagnostic detail is available to operators
        via server logs keyed by the X-Request-ID header.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ListAccountsResponse:
      type: object
      required:
        - accounts
        - pagination
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/Account"
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    ListJournalEntriesResponse:
      type: object
      required: [entries, pagination]
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/JournalEntry" }
        pagination: { $ref: "#/components/schemas/Pagination" }
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string